+++ src/gui/IconDownloader.cpp
@@ -17,7 +17,7 @@
 
 #include "IconDownloader.h"
 #include "core/Config.h"
-#include "core/NetworkManager.h"
+#include "networking/NetworkManager.h"
 
 #include <QBuffer>
 #include <QHostInfo>
