+++ src/networking/NetworkManager.cpp
